feat(mespapiers): Use new Steps dialog component to StepperDialogWrapper
Remove useless components
